feat: resolve inconsistent login state
In production when switching between computers, the local token would
make you believe you were logged in and trigger requests that were
guaranteed to fail. Instead of assuming the local token is correct we
let the server verify it by giving check the user id we get back from
the user locals call.

Signed-off-by: Alexander Alemayhu <alexander@alemayhu.com>